\name{motifCountMatrix}
\alias{motifCountMatrix}
\title{
Motif count matrix.}
\description{
The function can be used to retrieve a subset of the motif count matrix stored in a promoter.DB object, for a set of genes given as input. The promoter.DB objets are as generated by the functions makePromoterDBfromFimo or makePromoterDBfromRE.  }
\usage{
motifCountMatrix(Genes, promoterDB = "promoter.DB", strand = "both")
}
\arguments{
  \item{Genes}{
A list of the genes of interest.}
  \item{promoterDB}{
Promoter.DB object, }
  \item{strand}{
A character string that can take value: "plus", "minus" or "both". }
}
\details{
}
\value{
A promoter.DB object}
\references{
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
