\name{TFbindR-package}
\alias{TFbindR-package}
\alias{TFbindR}
\docType{package}
\title{
\packageTitle{TFbindR}
}
\description{
\packageDescription{TFbindR}
}
\details{

%The DESCRIPTION file:
%\packageDESCRIPTION{TFbindR}
\packageIndices{TFbindR}
}
\author{
Maja Ilievska 
Omid Safronov
Jarkko Salojärvi 

Maintainer: \packageMaintainer{TFbindR}
}

